fix: 팀 삭제 관련 문제 해결
* fix: team 삭제 연관관계 문제 해결

* fix: 팀매니저가 두 명 이상이라면 팀을 삭제하지 않는다

* fix: 테스트 코드 수정

* fix: 매니저 탈퇴 동시성 문제 해결

* refactor: imageUrl TEXT로 변경

* fix: 인코딩 깨짐 대비 프롬프트 수정

* fix: 중복된 메서드 삭제

* fix: TeamRecommendation 정리 추가

+++ src/main/java/site/codemonster/comon/domain/auth/service/MemberService.java
@@ -79,12 +79,15 @@ public class MemberService {
         teamApplyLowService.deleteTeamAppliesByMemberId(memberId);
 
         List<Team> teamsManagedByMember = teamLowService.findByTeamManagerId(memberId);
+
         for (Team team : teamsManagedByMember) {
             Long teamId = team.getTeamId();
 
-            articleLowService.deleteByTeamTeamId(teamId);
+            List<TeamMember> teamManagers = teamMemberLowService.findTeamManagerByTeamIdForUpdate(teamId);
 
-            teamMemberLowService.deleteByTeamTeamId(teamId);
+            if (teamManagers.size() > 1) continue;
+
+            articleLowService.deleteByTeamId(teamId);
 
             teamLowService.deleteById(teamId);
         }
